prevent from requesting voice instructions if the cache has been previously closed and add a check to delete the cache if there are files

+++ libandroid-navigation-ui/src/main/java/com/mapbox/services/android/navigation/ui/v5/voice/VoiceInstructionLoader.java
@@ -82,16 +82,20 @@ public class VoiceInstructionLoader {
   }
 
   void requestInstruction(String instruction, String textType, Callback<ResponseBody> callback) {
-    MapboxSpeech mapboxSpeech = mapboxSpeechBuilder
-      .instruction(instruction)
-      .textType(textType)
-      .build();
-    mapboxSpeech.enqueueCall(callback);
+    if (!cache.isClosed()) {
+      MapboxSpeech mapboxSpeech = mapboxSpeechBuilder
+        .instruction(instruction)
+        .textType(textType)
+        .build();
+      mapboxSpeech.enqueueCall(callback);
+    }
   }
 
   void flushCache() {
     try {
-      cache.delete();
+      if (cache.directory().listFiles() != null) {
+        cache.delete();
+      }
     } catch (IOException exception) {
       Timber.e(exception);
     }
